+++ internal/proxy/task.go
@@ -344,48 +344,6 @@ func (t *createCollectionTask) validateClusteringKey(ctx context.Context) error 
 	return nil
 }
 
-func (t *createCollectionTask) handleNamespaceProperty() error {
-	exists := hasNamespaceField(t.schema)
-	if exists {
-		return merr.WrapErrCollectionIllegalSchema(t.schema.Name,
-			"namespace field name is reserved and cannot be user-defined")
-	}
-
-	hasIsolation := hasIsolationProperty(t.GetProperties()...)
-	hasPartitionKey := hasPartitionKeyModeField(t.schema)
-	enabled, has, err := parseNamespaceProp(t.GetProperties()...)
-	if err != nil {
-		return err
-	}
-
-	if !has || !enabled {
-		return nil
-	}
-
-	if hasIsolation {
-		iso, err := common.IsPartitionKeyIsolationKvEnabled(t.GetProperties()...)
-		if err != nil {
-			return err
-		}
-		if !iso {
-			return merr.WrapErrCollectionIllegalSchema(t.schema.Name,
-				"isolation property is false when namespace enabled")
-		}
-	}
-
-	if hasPartitionKey {
-		return merr.WrapErrParameterInvalidMsg("namespace is not supported with partition key mode")
-	}
-
-	addNamespaceField(t.schema)
-	addIsolationProperty(t.schema)
-	log.Ctx(t.TraceCtx()).Info("added namespace field",
-		zap.String("collectionName", t.schema.Name),
-		zap.String("fieldName", common.NamespaceFieldName))
-
-	return nil
-}
-
 func (t *createCollectionTask) PreExecute(ctx context.Context) error {
 	t.Base.MsgType = commonpb.MsgType_CreateCollection
 	t.Base.SourceID = paramtable.GetNodeID()
@@ -446,10 +404,6 @@ func (t *createCollectionTask) PreExecute(ctx context.Context) error {
 
 	// validate field type definition
 	if err := validateFieldType(t.schema); err != nil {
-		return err
-	}
-
-	if err := t.handleNamespaceProperty(); err != nil {
 		return err
 	}
 
@@ -587,7 +541,7 @@ func (t *addCollectionFieldTask) PreExecute(ctx context.Context) error {
 	if typeutil.IsVectorType(t.fieldSchema.DataType) {
 		return merr.WrapErrParameterInvalidMsg(fmt.Sprintf("not support to add vector field, field name = %s", t.fieldSchema.Name))
 	}
-	if funcutil.SliceContain([]string{common.RowIDFieldName, common.TimeStampFieldName, common.MetaFieldName}, t.fieldSchema.GetName()) {
+	if funcutil.SliceContain([]string{common.RowIDFieldName, common.TimeStampFieldName, common.MetaFieldName, common.NamespaceFieldName}, t.fieldSchema.GetName()) {
 		return merr.WrapErrParameterInvalidMsg(fmt.Sprintf("not support to add system field, field name = %s", t.fieldSchema.Name))
 	}
 	if t.fieldSchema.IsPrimaryKey {
@@ -1164,63 +1118,11 @@ func hasNamespaceField(schema *schemapb.CollectionSchema) bool {
 	return false
 }
 
-func hasIsolationProperty(props ...*commonpb.KeyValuePair) bool {
-	for _, p := range props {
-		if p.GetKey() == common.PartitionKeyIsolationKey {
-			return true
-		}
-	}
-	return false
-}
-
-func addNamespaceField(schema *schemapb.CollectionSchema) {
-	fid := int64(-1)
-	for _, f := range schema.Fields {
-		if f.FieldID > fid {
-			fid = f.FieldID
-		}
-	}
-
-	for _, f := range schema.StructArrayFields {
-		if f.FieldID > fid {
-			fid = f.FieldID
-		}
-		for _, inner := range f.Fields {
-			if inner.FieldID > fid {
-				fid = inner.FieldID
-			}
-		}
-	}
-
-	schema.Fields = append(schema.Fields, &schemapb.FieldSchema{
-		FieldID:        fid + 1,
-		Name:           common.NamespaceFieldName,
-		IsPartitionKey: true,
-		DataType:       schemapb.DataType_VarChar,
-		TypeParams: []*commonpb.KeyValuePair{
-			{Key: common.MaxLengthKey, Value: fmt.Sprintf("%d", paramtable.Get().ProxyCfg.MaxVarCharLength.GetAsInt())},
-		},
-	})
-}
-
 func addIsolationProperty(schema *schemapb.CollectionSchema) {
 	schema.Properties = append(schema.Properties, &commonpb.KeyValuePair{
 		Key:   common.PartitionKeyIsolationKey,
 		Value: "true",
 	})
-}
-
-func parseNamespaceProp(props ...*commonpb.KeyValuePair) (value bool, has bool, err error) {
-	for _, p := range props {
-		if p.GetKey() == common.NamespaceEnabledKey {
-			value, err := strconv.ParseBool(p.GetValue())
-			if err != nil {
-				return false, false, merr.WrapErrParameterInvalidMsg(fmt.Sprintf("invalid namespace prop value: %s", p.GetValue()))
-			}
-			return value, true, nil
-		}
-	}
-	return false, false, nil
 }
 
 func hasMmapProp(props ...*commonpb.KeyValuePair) bool {
